Make parseVisit throw in case of an error
By making Main.parseVisit throw an exception in case an error occurs in
parsing, the following benefits can be achieved:

1. Do not exit the runtime in otherwise well-encapsulated methods
Main.parseVisit would have been free of severe side effects, if it would
not call System.exit() in case an error occurs in parsing.
Main.main() is now the only place that calls System.exit().

2. Do not leak to some otherwise unassociated class member
Main.parsingError is not a property of main, and scoped wrong by
definition. It is much better suited as a local variable, or even
encapsulated in the parsing process.
By writing a small implementation of ErrorListener we can encapsulate
that state nicely and query it after parsing has finished.

3. Unit tests now can test for the error case without checking for exit

4. As a consqeuence of (3.) we can get rid of a dependecy

Along the way:
 * Use JUnit's @Ignore instead of blindly commenting out test cases.
 * Fix error message talking about modules. We're not using modules.

Debatable: Main.parseVisit() now throws an ANTLR exception, this might
be considered a lack of abstraction.